+++ lib/pages/main_page.dart
@@ -2,7 +2,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:game_lists/database/database.dart';
 import 'package:game_lists/pages/main_page/games_widget_option.dart';
-import 'package:game_lists/pages/main_page/home_widget_option.dart';
 import 'package:game_lists/pages/main_page/search_widget_option.dart';
 import 'package:game_lists/pages/main_page/statistics_widget_option.dart';
 import 'package:game_lists/pages/main_page/widget_option.dart';
@@ -16,35 +15,51 @@ class MainPage extends StatefulWidget {
   _MainPageState createState() => _MainPageState();
 }
 
-class _MainPageState extends State<MainPage> {
+class _MainPageState extends State<MainPage> with TickerProviderStateMixin {
   int _selectedIndex = 0;
 
-  final List<WidgetOption> _widgetOptions = <WidgetOption>[
-    WidgetOption(
-      'Home',
-      title: 'Game List',
-      widget: HomeWidgetOption(),
-      iconData: Icons.home,
-    ),
-    WidgetOption(
-      'Search',
-      title: 'Search',
-      widget: SearchWidgetOption(),
-      iconData: Icons.search,
-    ),
-    WidgetOption(
-      'Games',
-      title: 'Games',
-      widget: GamesWidgetOption(),
-      iconData: Icons.games,
-    ),
-    WidgetOption(
-      'Statistics',
-      title: 'Statistics',
-      widget: StatisticsWidgetOption(),
-      iconData: Icons.insert_chart,
-    ),
-  ];
+  List<WidgetOption>? _widgetOptions;
+
+  @override
+  void initState() {
+    super.initState();
+    var gamesTabController = TabController(vsync: this, length: 5);
+    var statisticsTabController = TabController(vsync: this, length: 2);
+    _widgetOptions = <WidgetOption>[
+      WidgetOption(
+          title: 'Search',
+          iconData: Icons.search,
+          floatingActionButton: FloatingActionButton(
+              child: Icon(Icons.add),
+              onPressed: () async {
+                Navigator.pushNamed(context, '/adding_game');
+              }),
+          widget: SearchWidgetOption()),
+      WidgetOption(
+          title: 'Games',
+          tabController: gamesTabController,
+          isTabScrollable: true,
+          tabs: [
+            Tab(text: 'Inbox'),
+            Tab(text: 'Playing'),
+            Tab(text: 'Pause'),
+            Tab(text: 'Completed'),
+            Tab(text: 'Dropped')
+          ],
+          iconData: Icons.games,
+          widget: GamesWidgetOption(tabController: gamesTabController)),
+      WidgetOption(
+        title: 'Statistics',
+        tabController: statisticsTabController,
+        tabs: [
+          Tab(text: 'Player'),
+          Tab(text: 'Global'),
+        ],
+        iconData: Icons.insert_chart,
+        widget: StatisticsWidgetOption(tabController: statisticsTabController),
+      )
+    ];
+  }
 
   void _onItemTapped(int index) {
     setState(() {
@@ -54,83 +69,58 @@ class _MainPageState extends State<MainPage> {
 
   @override
   Widget build(BuildContext context) {
-    return DefaultTabController(
-      length: 5,
-      child: Scaffold(
-        floatingActionButton: _selectedIndex == 1
-            ? FloatingActionButton(
-                child: Icon(Icons.add),
-                onPressed: () async {
-                  Navigator.pushNamed(context, '/adding_game');
-                })
-            : null,
-        appBar: AppBar(
-          elevation: 0,
-          title: Text(_widgetOptions.elementAt(_selectedIndex).title),
-          bottom: _selectedIndex == 2
-              ? TabBar(
-                  isScrollable: true,
-                  tabs: [
-                    Tab(text: 'Inbox'),
-                    Tab(text: 'Playing'),
-                    Tab(text: 'Pause'),
-                    Tab(text: 'Completed'),
-                    Tab(text: 'Dropped')
-                  ],
-                )
-              : null,
-        ),
-        bottomNavigationBar: BottomNavigationBar(
-          items: _widgetOptions
-              .map((e) => BottomNavigationBarItem(icon: e.icon, label: e.label))
-              .toList(),
-          currentIndex: _selectedIndex,
-          type: BottomNavigationBarType.fixed,
-          selectedFontSize: 12,
-          onTap: _onItemTapped,
-          elevation: 0,
-        ),
-        drawer: Drawer(
-          child: ListView(
-            padding: EdgeInsets.zero,
-            children: <Widget>[
-              DrawerHeader(
-                child: Row(
-                  children: [
-                    Text(
-                      'Game Lists',
-                      style: TextStyle(fontSize: 24),
-                    ),
-                  ],
-                ),
-                decoration: BoxDecoration(
-                  color: Colors.red,
-                ),
+    return Scaffold(
+      floatingActionButton:
+          _widgetOptions!.elementAt(_selectedIndex).floatingActionButton,
+      appBar: _widgetOptions!.elementAt(_selectedIndex).appBar,
+      bottomNavigationBar: BottomNavigationBar(
+        items: _widgetOptions!.map((e) => e.bottomNavigationBarItem).toList(),
+        currentIndex: _selectedIndex,
+        type: BottomNavigationBarType.fixed,
+        selectedFontSize: 12,
+        onTap: _onItemTapped,
+        elevation: 0,
+      ),
+      drawer: Drawer(
+        child: ListView(
+          padding: EdgeInsets.zero,
+          children: <Widget>[
+            DrawerHeader(
+              child: Row(
+                children: [
+                  Text(
+                    'Game Lists',
+                    style: TextStyle(fontSize: 24),
+                  ),
+                ],
               ),
-              ListTile(
-                title: Text('Settings'),
-                onTap: () {
-                  Navigator.pushNamed(context, '/settings');
-                },
+              decoration: BoxDecoration(
+                color: Colors.red,
               ),
-              ListTile(
-                title: Text('About'),
-                onTap: () {
-                  Navigator.pushNamed(context, '/about');
-                },
-              ),
-              ListTile(
-                title: Text('Exit'),
-                onTap: () {
-                  SystemNavigator.pop();
-                },
-              ),
-            ],
-          ),
+            ),
+            ListTile(
+              title: Text('Settings'),
+              onTap: () {
+                Navigator.pushNamed(context, '/settings');
+              },
+            ),
+            ListTile(
+              title: Text('About'),
+              onTap: () {
+                Navigator.pushNamed(context, '/about');
+              },
+            ),
+            ListTile(
+              title: Text('Exit'),
+              onTap: () {
+                SystemNavigator.pop();
+              },
+            ),
+          ],
         ),
-        body: Center(
-          child: _widgetOptions.elementAt(_selectedIndex).widget,
-        ),
+      ),
+      body: Center(
+        child: _widgetOptions!.elementAt(_selectedIndex).widget,
       ),
     );
   }