+++ lib/pages/main_page/games_widget_option.dart
@@ -1,9 +1,19 @@
 import 'package:flutter/material.dart';
 
-class GamesWidgetOption extends StatelessWidget {
+class GamesWidgetOption extends StatefulWidget {
+  GamesWidgetOption({Key? key, this.tabController}) : super(key: key);
+
+  @override
+  _GamesWidgetOptionState createState() => _GamesWidgetOptionState();
+
+  final TabController? tabController;
+}
+
+class _GamesWidgetOptionState extends State<GamesWidgetOption> {
   @override
   Widget build(BuildContext context) {
     return TabBarView(
+      controller: widget.tabController,
       children: [
         Icon(Icons.directions_car),
         Icon(Icons.directions_transit),
@@ -12,6 +22,5 @@ class GamesWidgetOption extends StatelessWidget {
         Icon(Icons.directions_bike),
       ],
     );
-    ;
   }
 }