+++ lib/pages/main_page/statistics_widget_option.dart
@@ -1,8 +1,23 @@
 import 'package:flutter/material.dart';
 
-class StatisticsWidgetOption extends StatelessWidget {
+class StatisticsWidgetOption extends StatefulWidget {
+  StatisticsWidgetOption({Key? key, this.tabController}) : super(key: key);
+
+  @override
+  _StatisticsWidgetOptionState createState() => _StatisticsWidgetOptionState();
+
+  final TabController? tabController;
+}
+
+class _StatisticsWidgetOptionState extends State<StatisticsWidgetOption> {
   @override
   Widget build(BuildContext context) {
-    return Text('Statistics');
+    return TabBarView(
+      controller: widget.tabController,
+      children: [
+        Icon(Icons.directions_car),
+        Icon(Icons.directions_transit),
+      ],
+    );
   }
 }