+++ lib/pages/main_page/widget_option.dart
@@ -1,12 +1,29 @@
 import 'package:flutter/material.dart';
 
 class WidgetOption {
-  final String title;
-  final String label;
+  final TabController? tabController;
+  final AppBar appBar;
+  final BottomNavigationBarItem bottomNavigationBarItem;
+  final FloatingActionButton? floatingActionButton;
   final Widget widget;
-  final Icon icon;
 
-  WidgetOption(this.label,
-      {required this.title, required this.widget, required IconData iconData})
-      : icon = Icon(iconData);
+  WidgetOption(
+      {
+        required String title,
+        this.tabController,
+        bool? isTabScrollable,
+        List<Tab>? tabs,
+        this.floatingActionButton,
+        required IconData iconData,
+        required this.widget,
+
+      }) : appBar = AppBar(
+        elevation: 0,
+        title: Text(title),
+        bottom: tabs != null ? TabBar(
+            controller: tabController,
+            isScrollable: isTabScrollable ?? false,
+            tabs: tabs
+        ): null
+  ), bottomNavigationBarItem = BottomNavigationBarItem(icon: Icon(iconData), label: title);
 }