+++ lib/game_lists.dart
@@ -2,9 +2,14 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:game_lists/database/database.dart';
 import 'package:game_lists/pages/about_page.dart';
+import 'package:game_lists/pages/country_page.dart';
+import 'package:game_lists/pages/developer_page.dart';
 import 'package:game_lists/pages/game_page.dart';
 import 'package:game_lists/pages/main_page.dart';
+import 'package:game_lists/pages/platform_page.dart';
+import 'package:game_lists/pages/series_page.dart';
 import 'package:game_lists/pages/settings_page.dart';
+import 'package:game_lists/pages/tag_page.dart';
 
 class GameLists extends StatefulWidget {
   const GameLists({Key? key, this.database}) : super(key: key);
@@ -27,9 +32,14 @@ class _GameListsState extends State<GameLists> {
       initialRoute: '/',
       routes: {
         '/': (context) => MainPage(database: widget.database),
-        '/settings': (context) => SettingsPage(),
         '/about': (context) => AboutPage(),
-        '/game': (context) => GamePage()
+        '/country': (context) => CountryPage(),
+        '/developer': (context) => DeveloperPage(),
+        '/game': (context) => GamePage(),
+        '/platform': (context) => PlatformPage(),
+        '/series': (context) => SeriesPage(),
+        '/settings': (context) => SettingsPage(),
+        '/tag': (context) => TagPage(),
       },
       debugShowCheckedModeBanner: false,
     );