+++ lib/pages/country_page.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+
+class CountryPage extends StatefulWidget {
+  @override
+  _CountryPageState createState() => _CountryPageState();
+}
+
+class _CountryPageState extends State<CountryPage> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Country'),
+      ),
+      body: Container(),
+    );
+  }
+}