+++ lib/pages/platform_page.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+
+class PlatformPage extends StatefulWidget {
+  @override
+  _PlatformPageState createState() => _PlatformPageState();
+}
+
+class _PlatformPageState extends State<PlatformPage> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Platform'),
+      ),
+      body: Container(),
+    );
+  }
+}