+++ lib/pages/tag_page.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+
+class TagPage extends StatefulWidget {
+  @override
+  _TagPageState createState() => _TagPageState();
+}
+
+class _TagPageState extends State<TagPage> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Tag'),
+      ),
+      body: Container(),
+    );
+  }
+}